% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readCNAprobesegments}
\alias{readCNAprobesegments}
\title{Read CNA data with probe and segmentation files}
\usage{
readCNAprobesegments(
  wd,
  probefilename = "probes.txt",
  segmentfilename = "segments.txt",
  sub = ".final",
  ...
)
}
\description{
Read CNA data with probe and segmentation files
}
