% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readMultianno}
\alias{readMultianno}
\title{Read multianno annotated files with nested folder structure}
\usage{
readMultianno(
  wd,
  folderprefix = "^R",
  filesuffix = "_multianno.txt",
  sep = "\\t",
  header = TRUE,
  na.strings = ".",
  ...
)
}
\description{
Read multianno annotated files with nested folder structure
}
